feat: unify role and permission requests into access_requests
Replace the separate role_requests and permission_requests tables with
a single access_requests table that bundles a Zitadel role request
with fine-grained OpenFGA permissions in a JSONB column.

On approve: Zitadel role is granted + all OpenFGA tuples are created.
On revoke: Zitadel role is removed + all OpenFGA tuples are deleted.

New files:
  - AccessRequestRepository: unified repository with JSONB permissions
  - AccessRequestHandler: user-facing submit + list + status endpoints
  - AccessRequestAdminHandler: admin approve/reject/revoke with
    combined Zitadel role sync + OpenFGA tuple management

Removed:
  - RoleRequestRepository, RoleRequestHandler, RoleRequestAdminHandler
  - PermissionRequestRepository, PermissionRequestHandler,
    PermissionRequestAdminHandler

Routes changed:
  - /auth/role-requests + /auth/permission-requests → /auth/access-requests
  - /admin/role-requests + /admin/permission-requests → /admin/access-requests

Refs: Liturgical-Calendar/LiturgicalCalendarAPI#517
