+++ infrastructure/init-db.sql
@@ -30,73 +30,45 @@ GRANT ALL ON SCHEMA public TO litcal;
 -- This runs as postgres superuser (which has permissions), but in the litcal database context
 CREATE EXTENSION IF NOT EXISTS pgcrypto;
 
--- Role requests table
--- Stores pending role assignment requests from users
-CREATE TABLE role_requests (
+-- Unified access requests table
+-- Bundles a Zitadel role request with fine-grained OpenFGA permissions.
+-- On approval: Zitadel role is granted + all OpenFGA tuples are created.
+-- On revoke: Zitadel role is removed + all OpenFGA tuples are deleted.
+--
+-- Granted calendar permissions are managed by OpenFGA relationship tuples.
+-- See infrastructure/openfga-model.json and scripts/setup-openfga.sh.
+CREATE TABLE access_requests (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     zitadel_user_id VARCHAR(255) NOT NULL,
     user_email VARCHAR(255) NOT NULL,
     user_name VARCHAR(255),
-    requested_role VARCHAR(50) NOT NULL,  -- 'developer', 'calendar_editor', 'test_editor'
+    requested_role VARCHAR(50) NOT NULL,
+    permissions JSONB NOT NULL DEFAULT '[]',
     justification TEXT,
+    credentials TEXT,
     status VARCHAR(20) DEFAULT 'pending',
-    reviewed_by VARCHAR(255),              -- Admin's Zitadel user ID
+    reviewed_by VARCHAR(255),
     review_notes TEXT,
     zitadel_sync_status VARCHAR(20) DEFAULT NULL,
     zitadel_sync_error TEXT DEFAULT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     reviewed_at TIMESTAMP,
     CONSTRAINT chk_requested_role CHECK (requested_role IN ('developer', 'calendar_editor', 'test_editor')),
-    CONSTRAINT chk_role_request_status CHECK (status IN ('pending', 'approved', 'rejected', 'revoked')),
+    CONSTRAINT chk_access_request_status CHECK (status IN ('pending', 'approved', 'rejected', 'revoked')),
     CONSTRAINT chk_zitadel_sync_status CHECK (zitadel_sync_status IS NULL OR zitadel_sync_status IN ('pending', 'synced', 'failed'))
 );
 
-CREATE INDEX idx_role_requests_status ON role_requests(status);
-CREATE INDEX idx_role_requests_user ON role_requests(zitadel_user_id);
-CREATE INDEX idx_role_requests_created ON role_requests(created_at);
-CREATE INDEX idx_role_requests_sync_status ON role_requests(zitadel_sync_status)
+CREATE INDEX idx_access_requests_status ON access_requests(status);
+CREATE INDEX idx_access_requests_user ON access_requests(zitadel_user_id);
+CREATE INDEX idx_access_requests_created ON access_requests(created_at);
+CREATE INDEX idx_access_requests_sync_status ON access_requests(zitadel_sync_status)
 WHERE zitadel_sync_status = 'failed';
 
-COMMENT ON TABLE role_requests IS 'Pending role assignment requests from users';
-COMMENT ON COLUMN role_requests.requested_role IS 'Role: developer, calendar_editor, test_editor';
-COMMENT ON COLUMN role_requests.status IS 'Status: pending, approved, rejected, revoked';
-
--- NOTE: Granted calendar permissions (formerly user_calendar_permissions) are now
--- managed by OpenFGA relationship tuples. See infrastructure/openfga-model.json
--- and scripts/setup-openfga.sh for the fine-grained authorization model.
-
--- Permission requests (approval workflow)
--- Users request access to specific calendars; admins approve/reject.
--- On approval, an OpenFGA tuple is created via the admin permissions endpoint.
-CREATE TABLE permission_requests (
-    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    zitadel_user_id VARCHAR(255) NOT NULL,
-    user_email VARCHAR(255) NOT NULL,
-    user_name VARCHAR(255),
-    object_type VARCHAR(30) NOT NULL,
-    object_id VARCHAR(50) NOT NULL,
-    relation VARCHAR(20) NOT NULL,
-    justification TEXT,
-    credentials TEXT,
-    status VARCHAR(20) DEFAULT 'pending',
-    reviewed_by VARCHAR(255),
-    review_notes TEXT,
-    created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    reviewed_at TIMESTAMP,
-    CONSTRAINT chk_permission_requests_status CHECK (status IN ('pending', 'approved', 'rejected', 'revoked')),
-    CONSTRAINT chk_permission_requests_object_type CHECK (object_type IN ('national_calendar', 'diocesan_calendar', 'wider_region', 'test_definition')),
-    CONSTRAINT chk_permission_requests_relation CHECK (relation IN ('admin', 'viewer', 'editor', 'deleter'))
-);
-
-CREATE INDEX idx_permission_requests_status ON permission_requests(status);
-CREATE INDEX idx_permission_requests_user ON permission_requests(zitadel_user_id);
-CREATE INDEX idx_permission_requests_object ON permission_requests(object_type, object_id);
-
-COMMENT ON TABLE permission_requests IS 'Approval workflow for calendar access — on approval, OpenFGA tuple is created';
-COMMENT ON COLUMN permission_requests.object_type IS 'OpenFGA object type: national_calendar, diocesan_calendar, wider_region, test_definition';
-COMMENT ON COLUMN permission_requests.object_id IS 'Resource identifier: USA, BOSTON, Americas, etc.';
-COMMENT ON COLUMN permission_requests.relation IS 'OpenFGA relation: admin, viewer, editor, deleter';
-COMMENT ON COLUMN permission_requests.status IS 'Status: pending, approved, rejected, revoked';
+COMMENT ON TABLE access_requests IS 'Unified role + permission requests — role via Zitadel, permissions via OpenFGA';
+COMMENT ON COLUMN access_requests.requested_role IS 'Zitadel role: developer, calendar_editor, test_editor';
+COMMENT ON COLUMN access_requests.permissions IS 'JSON array of OpenFGA tuples: [{object_type, object_id, relation}, ...]';
+COMMENT ON COLUMN access_requests.status IS 'Status: pending, approved, rejected, revoked';
+COMMENT ON COLUMN access_requests.zitadel_sync_status IS 'Zitadel role sync: null (not attempted), pending, synced, failed';
 
 -- Applications (for API developers)
 CREATE TABLE applications (
