+++ src/Handlers/Admin/NotificationsHandler.php
@@ -12,9 +12,8 @@ use LiturgicalCalendar\Api\Http\Enum\RequestMethod;
 use LiturgicalCalendar\Api\Http\Exception\ForbiddenException;
 use LiturgicalCalendar\Api\Http\Exception\UnauthorizedException;
 use LiturgicalCalendar\Api\Http\Middleware\OidcAuthMiddleware;
+use LiturgicalCalendar\Api\Repositories\AccessRequestRepository;
 use LiturgicalCalendar\Api\Repositories\ApplicationRepository;
-use LiturgicalCalendar\Api\Repositories\PermissionRequestRepository;
-use LiturgicalCalendar\Api\Repositories\RoleRequestRepository;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -28,9 +27,8 @@ use Psr\Http\Message\ServerRequestInterface;
  */
 final class NotificationsHandler extends AbstractHandler
 {
-    private ?RoleRequestRepository $roleRequestRepo       = null;
-    private ?ApplicationRepository $applicationRepo       = null;
-    private ?PermissionRequestRepository $permRequestRepo = null;
+    private ?AccessRequestRepository $accessRequestRepo = null;
+    private ?ApplicationRepository $applicationRepo     = null;
 
     public function __construct()
     {
@@ -41,12 +39,12 @@ final class NotificationsHandler extends AbstractHandler
         $this->allowCredentials      = true;
     }
 
-    private function getRoleRequestRepository(): RoleRequestRepository
+    private function getAccessRequestRepository(): AccessRequestRepository
     {
-        if ($this->roleRequestRepo === null) {
-            $this->roleRequestRepo = new RoleRequestRepository();
+        if ($this->accessRequestRepo === null) {
+            $this->accessRequestRepo = new AccessRequestRepository();
         }
-        return $this->roleRequestRepo;
+        return $this->accessRequestRepo;
     }
 
     private function getApplicationRepository(): ApplicationRepository
@@ -55,14 +53,6 @@ final class NotificationsHandler extends AbstractHandler
             $this->applicationRepo = new ApplicationRepository();
         }
         return $this->applicationRepo;
-    }
-
-    private function getPermissionRequestRepository(): PermissionRequestRepository
-    {
-        if ($this->permRequestRepo === null) {
-            $this->permRequestRepo = new PermissionRequestRepository();
-        }
-        return $this->permRequestRepo;
     }
 
     public function handle(ServerRequestInterface $request): ResponseInterface
@@ -95,31 +85,28 @@ final class NotificationsHandler extends AbstractHandler
 
         // Get notification counts
         $notifications = [
-            'pending_role_requests'       => 0,
-            'pending_applications'        => 0,
-            'pending_permission_requests' => 0,
-            'total'                       => 0,
-            'items'                       => [],
+            'pending_access_requests' => 0,
+            'pending_applications'    => 0,
+            'total'                   => 0,
+            'items'                   => [],
         ];
 
         if (Connection::isConfigured()) {
-            // Get role request counts
-            $roleRequestRepo = $this->getRoleRequestRepository();
-            $counts          = $roleRequestRepo->getRequestCounts();
+            // Get access request counts (unified role + permission requests)
+            $accessRequestRepo                        = $this->getAccessRequestRepository();
+            $notifications['pending_access_requests'] = $accessRequestRepo->countPending();
 
-            $notifications['pending_role_requests'] = $counts['pending'];
-
-            // Get recent pending role requests for the dropdown
-            $pendingRequests = $roleRequestRepo->getPendingRequests();
+            // Get recent pending access requests for the dropdown
+            $pendingRequests = $accessRequestRepo->getPending();
             foreach (array_slice($pendingRequests, 0, 5) as $req) {
                 $notifications['items'][] = [
-                    'type'       => 'role_request',
+                    'type'       => 'access_request',
                     'id'         => $req['id'] ?? '',
                     'user_name'  => $req['user_name'] ?? $req['user_email'] ?? 'Unknown',
                     'user_email' => $req['user_email'] ?? '',
                     'role'       => $req['requested_role'] ?? '',
                     'created_at' => $req['created_at'] ?? '',
-                    'url'        => 'admin-role-requests.php',
+                    'url'        => 'admin-permissions.php',
                 ];
             }
 
@@ -141,26 +128,6 @@ final class NotificationsHandler extends AbstractHandler
                 ];
             }
 
-            // Get pending permission requests count
-            $permRequestRepo                              = $this->getPermissionRequestRepository();
-            $notifications['pending_permission_requests'] = $permRequestRepo->countPending();
-
-            // Get recent pending permission requests for the dropdown
-            $pendingPermRequests = $permRequestRepo->getPending();
-            foreach (array_slice($pendingPermRequests, 0, 5) as $req) {
-                $notifications['items'][] = [
-                    'type'        => 'permission_request',
-                    'id'          => $req['id'] ?? '',
-                    'user_name'   => $req['user_name'] ?? $req['user_email'] ?? 'Unknown',
-                    'user_email'  => $req['user_email'] ?? '',
-                    'object_type' => $req['object_type'] ?? '',
-                    'object_id'   => $req['object_id'] ?? '',
-                    'relation'    => $req['relation'] ?? '',
-                    'created_at'  => $req['created_at'] ?? '',
-                    'url'         => 'admin-permissions.php',
-                ];
-            }
-
             // Sort items by created_at descending and limit to 5 most recent
             usort($notifications['items'], function ($a, $b) {
                 $aDate = is_string($a['created_at']) ? $a['created_at'] : '';
@@ -169,9 +136,8 @@ final class NotificationsHandler extends AbstractHandler
             });
             $notifications['items'] = array_slice($notifications['items'], 0, 5);
 
-            $notifications['total'] = $notifications['pending_role_requests']
-                                    + $notifications['pending_applications']
-                                    + $notifications['pending_permission_requests'];
+            $notifications['total'] = $notifications['pending_access_requests']
+                                    + $notifications['pending_applications'];
         }
 
         // Add Cache-Control header to prevent caching
