+++ src/Router.php
@@ -23,14 +23,12 @@ use LiturgicalCalendar\Api\Handlers\Auth\LoginHandler;
 use LiturgicalCalendar\Api\Handlers\Auth\LogoutHandler;
 use LiturgicalCalendar\Api\Handlers\Auth\MeHandler;
 use LiturgicalCalendar\Api\Handlers\Auth\RefreshHandler;
-use LiturgicalCalendar\Api\Handlers\Auth\RoleRequestHandler;
+use LiturgicalCalendar\Api\Handlers\Auth\AccessRequestHandler;
 use LiturgicalCalendar\Api\Handlers\Auth\EmailVerificationHandler;
-use LiturgicalCalendar\Api\Handlers\Auth\PermissionRequestHandler;
+use LiturgicalCalendar\Api\Handlers\Admin\AccessRequestAdminHandler;
 use LiturgicalCalendar\Api\Handlers\Admin\ApplicationAdminHandler;
 use LiturgicalCalendar\Api\Handlers\Admin\NotificationsHandler;
 use LiturgicalCalendar\Api\Handlers\Admin\PermissionAdminHandler;
-use LiturgicalCalendar\Api\Handlers\Admin\PermissionRequestAdminHandler;
-use LiturgicalCalendar\Api\Handlers\Admin\RoleRequestAdminHandler;
 use LiturgicalCalendar\Api\Handlers\Admin\UsersHandler;
 use LiturgicalCalendar\Api\Handlers\ApplicationsHandler;
 use LiturgicalCalendar\Api\Http\Enum\StatusCode;
@@ -340,19 +338,13 @@ class Router
                     } elseif ($authRoute === 'me') {
                         $meHandler     = new MeHandler();
                         $this->handler = $meHandler;
-                    } elseif ($authRoute === 'role-requests') {
-                        // Role request routes for authenticated users
-                        // POST /auth/role-requests - Create new request
-                        // GET /auth/role-requests - Get user's own requests
-                        // GET /auth/role-requests/status - Check if user needs to request a role
-                        $roleRequestHandler = new RoleRequestHandler();
-                        $this->handler      = $roleRequestHandler;
-                    } elseif ($authRoute === 'permission-requests') {
-                        // Permission request routes for authenticated users
-                        // POST /auth/permission-requests - Submit a new permission request
-                        // GET /auth/permission-requests - View own requests
-                        $permissionRequestHandler = new PermissionRequestHandler();
-                        $this->handler            = $permissionRequestHandler;
+                    } elseif ($authRoute === 'access-requests') {
+                        // Unified access request routes for authenticated users
+                        // POST /auth/access-requests - Submit access request (role + permissions)
+                        // GET /auth/access-requests - View own requests
+                        // GET /auth/access-requests/status - Check access status
+                        $accessRequestHandler = new AccessRequestHandler();
+                        $this->handler        = $accessRequestHandler;
                     } elseif ($authRoute === 'email-verification') {
                         // Email verification routes for authenticated users
                         // POST /auth/email-verification/resend - Resend verification email
@@ -371,13 +363,14 @@ class Router
                 // Handle admin routes
                 if (count($requestPathParts) >= 1) {
                     $adminRoute = $requestPathParts[0];
-                    if ($adminRoute === 'role-requests') {
-                        // Admin role request management routes
-                        // GET /admin/role-requests - List all pending requests
-                        // POST /admin/role-requests/{id}/approve - Approve a request
-                        // POST /admin/role-requests/{id}/reject - Reject a request
-                        $roleRequestAdminHandler = new RoleRequestAdminHandler();
-                        $this->handler           = $roleRequestAdminHandler;
+                    if ($adminRoute === 'access-requests') {
+                        // Unified access request management routes
+                        // GET  /admin/access-requests - List requests
+                        // POST /admin/access-requests/{id}/approve - Approve (role + tuples)
+                        // POST /admin/access-requests/{id}/reject - Reject
+                        // POST /admin/access-requests/{id}/revoke - Revoke (role + tuples)
+                        $accessRequestAdminHandler = new AccessRequestAdminHandler();
+                        $this->handler             = $accessRequestAdminHandler;
                     } elseif ($adminRoute === 'notifications') {
                         // Admin notifications route
                         // GET /admin/notifications - Get counts of pending items
@@ -389,14 +382,6 @@ class Router
                         // DELETE /admin/users/{userId}/roles/{role} - Revoke a role
                         $usersHandler  = new UsersHandler();
                         $this->handler = $usersHandler;
-                    } elseif ($adminRoute === 'permission-requests') {
-                        // Admin permission request review routes
-                        // GET  /admin/permission-requests - List pending requests
-                        // POST /admin/permission-requests/{id}/approve - Approve
-                        // POST /admin/permission-requests/{id}/reject - Reject
-                        // POST /admin/permission-requests/{id}/revoke - Revoke
-                        $permRequestAdminHandler = new PermissionRequestAdminHandler();
-                        $this->handler           = $permRequestAdminHandler;
                     } elseif ($adminRoute === 'permissions') {
                         // Admin permission management routes (OpenFGA)
                         // GET    /admin/permissions       - List permissions (with filters)
@@ -564,10 +549,10 @@ class Router
             $pipeline->pipe(new HttpsEnforcementMiddleware());
         }
 
-        // Apply OIDC authentication for auth routes (role-requests, permission-requests, email-verification), admin, and applications
+        // Apply OIDC authentication for auth routes (access-requests, email-verification), admin, and applications
         // These routes need the oidc_user attribute set before the handler checks authentication
         if (
-            ( $route === 'auth' && count($requestPathParts) >= 1 && in_array($requestPathParts[0], ['role-requests', 'permission-requests', 'email-verification'], true) )
+            ( $route === 'auth' && count($requestPathParts) >= 1 && in_array($requestPathParts[0], ['access-requests', 'email-verification'], true) )
             || $route === 'admin'
             || $route === 'applications'
         ) {
